requirements.txt
django
djangorestframework
markdown
django-filter
mysqlclient
django-cors-headers
django-url-filter
